Invert schmitt trigger symbol for more common convention
Schmitt triggers are usually depicted with a triangular symbol containing a smaller element representing hysterisis. There are two ways to depict this symbol and they are used to represent inverting and non-inverting schitt triggers. There is no solid convention as to which is which, but it appears to be more common to depict a signal going from low to high for the non-inverting device. This is the convention I have chosen and I have corrected the image to match.