Add a few color sets (conventions) found online
I've adapted a few of the wire-color conventions listed in this
discussion:
<https://community.vcvrack.com/t/choosing-cable-colors-in-vcv-1-0/4155>

These will appear as options in WM-101 *if* there isn't already a
`WM-101.json` file for this user. The existing default colors will
then be active, just as before.

Ideally we'd have a feature to add/merge these shared color sets with
the user's saved settings.