Working billboard view (shows labels, if any)
I'm using multiple instances of mirrored text to outline labels with
black, so they should work on any background.

I've tried to strike a balance between legibility and possibly long
labels. We can review this for legibility at typical zoom levels for
Youtube, etc. and consider a heavier font, and possibly the ability to
widen this module to show more.